BREAKING CHANGE: Totally refactor @valos/prophet
This commit begins the refactoring process.

This refactoring process will linearize the upstream/downstream from
the mongrel it is now, where Oracle has two upstreams: Scribe and
the authorities. The reordering swaps the order of Scribe and Oracle
and pushes the authorities to be the upstream of Oracle.

Fabric prophet/connection API's will undergo a breaking overhaul.

Commits following this commit are assumed to be broken until next
prerelease version increase.