BREAKING CHANGE(prophet): Begin complete @valos/prophet restructuring
This reordering process will linearize the upstream/downstream from
the mongrel it is now, where Oracle has two upstreams: Scribe and
the authorities. The reordering swaps the order of Scribe and Oracle
and pushes the authorities to be the upstream of Oracle.

Fabric prophet/connection API's will undergo a breaking overhaul.

Commits following this commit will be broken until next version
increase.