refactor(prophet): Change into plural events as receiveTruths/Commands
In practice also fully overhauls Scribe.
A lot of the prophet API are overhauled and more clearly defined.
Now downstream pathways can and must always accept batches of events,
as it always should have been. This simplifies performant designs a lot,
as no specific caching and pooling needs to be done for
narrate/chronicleEvents pathways.