feature: Add support for references and functions in Resource properties
This and previous work is seriously converging the differences between
KueryExpression, Literal and Identifier. Their Expression types and
field names are different (vakon, value and reference respectively)
but their serialization is identical.
Once the property rework is carried out to completion they should be
unified and the Expression Data indirection be removed.
This necessitates Zero and possibly other ValaaSpace changes, so
support must be in place for proper introspection via property
descriptors.