refactor: Renames transaction -> discourse, adds (Story|Passage)IndexTag
Adds Story
Previous naming was slightly confusing as sometimes an
options.transaction was not necessarily a transaction but
just a plain discourse. As all transaction API was identical
to discourse API anyway this naming change does not break
anything conceptually.
On one hand transaction semantics are very local as
acquireTransaction must be paired with releaseTransaction in
the syntactic environment, and on the other hand even
non-transactional discourses behave transaction-like in terms
of their command->truth evaluation hooks.